Fix test failures related to the addition of VolumeSizeGB.
[#113015649](https://www.pivotaltracker.com/story/show/113015649)

+++ src/light-stemcell-builder/driver/machine_image_lifecycle_test.go
@@ -109,7 +109,7 @@ var _ = Describe("Machine Image Lifecycle", func() {
 		Expect(resp.StatusCode).To(Equal(http.StatusOK))
 
 		Expect(m.FileFormat).To(Equal(imageFormat))
-		Expect(m.VolumeSizeGB).To(Equal(3))
+		Expect(m.VolumeSizeGB).To(Equal(int64(3)))
 
 		deleteDriver := driver.NewDeleteMachineImageDriver(GinkgoWriter, creds)
